Improvement and UI fix
Fixed bug where launching an app with the switcher open leaves the
switcher open (but ignored by the app).

Also blacked out SpringBoard while changing apps to avoid interacting
with it when launching big apps lags.

Also changed the NSString/SBapplication code to try and detect what it
is rather then just hard coding the iOS version.

Also upped version number to a bigger one the the one in Cydia (to
avoid the annoying prompting to 'update' to an older version.